Restructured creation from away from static and used a Builder
By removing the static requirements from nearly everything in this class, it
will now be possible to create an AppRater dialog that only exists in heap
for the life of one context. If a user still wants a static AppRater object then
they are free to implement that by creating a singleton object.

The Builder pattern was implemented as it is great for classes that tend
towards the telescoping constructor anti-pattern, which makes
maintainability and upgradability very difficult, and this is cropping up in the
note in pull request ticket #32. This will deal with the feeling "weary of
adding yet another boolean, the lib is getting a little complex/cluttered", and
is a "design to better allow the various options that are being added".